Fix `readability-redundant-casting` in `SimpleMergeSelector`
`cur_age` is already `size_t`, so the explicit cast made the arm_tidy build fail:
https://s3.amazonaws.com/clickhouse-test-reports/json.html?PR=100436&sha=dbd4921b1bf5083f362dc6e4b038e750bc3b0347&name_0=PR&name_1=Build%20%28arm_tidy%29
https://github.com/ClickHouse/ClickHouse/pull/100436

+++ src/Storages/MergeTree/Compaction/MergeSelectors/SimpleMergeSelector.cpp
@@ -354,7 +354,7 @@ void selectWithinPartsRange(
             max_age = std::max(max_age, cur_age);
             all_small_and_fresh = all_small_and_fresh
                 && cur_size < settings.small_parts_threshold
-                && static_cast<size_t>(cur_age) < settings.small_parts_max_age;
+                && cur_age < settings.small_parts_max_age;
 
             if (max_parts_to_merge_at_once && end - begin > max_parts_to_merge_at_once)
             {
